fix(arch): move ResultsContent to common/widgets to break cross-feature import (#1934)
EventAdmissionController imported ResultsContent directly from the home
feature's internal widget tree (home/widgets/event_now_phases/), creating
an event→home cross-feature dependency that violates feature-first isolation.

Resolution:
- New: common/widgets/match_results_content.dart with MatchResultsContent
- home/widgets/event_now_phases/results_content.dart re-exports via typedef
  so home's internal import path is unchanged
- event_admission_controller.dart now imports from common and uses MatchResultsContent

Confidence: high
Scope-risk: narrow
Rejected: inline ResultsContent body into controller | increases duplication
Rejected: move to event/widgets | home would then import from event (same problem)
